requirements.txt
streamlit
replicate
transformers
python-dotenv